+++ src/ShapeCrawler/ShapeCollection/CopyableShape.cs
@@ -17,6 +17,6 @@ internal abstract class CopyableShape : Shape
         P.ShapeTree pShapeTree,
         IEnumerable<string> existingShapeNames)
     {
-        new WrappedPShapeTree(pShapeTree).Add(this.pShapeTreeElement);
+        new WrappedPShapeTree(pShapeTree).Add(this.PShapeTreeElement);
     }
 }